+++ cli-generator/src/main/java/com/dmj/cli/domain/Product.java
@@ -74,11 +74,11 @@ public class Product extends BaseEntity implements Serializable {
 
     @ApiModelProperty(value = "浏览数")
     @TableField("view_num")
-    private Integer viewNum;
+    private Long viewNum;
 
     @ApiModelProperty(value = "点赞数")
     @TableField("favour_num")
-    private Integer favourNum;
+    private Long favourNum;
 
     @ApiModelProperty(value = "通过理由")
     @TableField("pass_reason")
